energy and water consumption telemetry is started.

+++ src/main/java/org/thingsboard/trendz/generator/solution/greenhouse/model/EnergyMeter.java
@@ -23,5 +23,5 @@ public class EnergyMeter extends ModelEntity {
     private String systemName;
     private String systemLabel;
 
-    private Telemetry<Integer> consumptionEnergy;
+    private Telemetry<Double> consumptionEnergy;
 }
